pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "academia-mcp"
version = "1.9.0"
description = "MCP server that provides different tools to search for scientific publications"
readme = "README.md"
authors = [
    {name = "Ilya Gusev", email = "phoenixilya@gmail.com"}
]
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "mcp>=1.10.1",
    "xmltodict>=0.14.0",
    "types-xmltodict>=0.14.0",
    "requests>=2.32.0",
    "types-requests>=2.32.0",
    "pypdf>=5.1.0",
    "beautifulsoup4>=4.12.0",
    "types-beautifulsoup4>=4.12.0",
    "markdownify==0.14.1",
    "acl-anthology==0.5.2",
    "markdown==3.7.0",
    "types-markdown==3.7.0.20250322",
    "huggingface-hub>=0.32.4",
    "fire>=0.7.0",
    "openai>=1.97.1",
    "jinja2>=3.1.6",
    "datasets>=4.0.0",
    "pymupdf>=1.26.4",
    "pillow>=11.3.0",
    "pydantic-settings>=2.6.0",
]

[dependency-groups]
dev = [
    "black==25.1.0",
    "mypy==1.16.0",
    "flake8==7.2.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
]

[project.urls]
Homepage = "https://github.com/IlyaGusev/academia_mcp"

[project.scripts]
academia_mcp = "academia_mcp:main"

[[tool.mypy.overrides]]
module = "*"
ignore_missing_imports = true
follow_imports = "skip"

[tool.setuptools.packages.find]
where = ["."]
include = ["academia_mcp*"]

[tool.setuptools.package-data]
"academia_mcp" = ["py.typed", "latex_templates/**/*"]
"academia_mcp.tools" = ["py.typed"]

[[tool.mypy.overrides]]
module = "pydantic.*"
follow_imports = "skip"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_test_loop_scope = "function"
filterwarnings = [
    "ignore:builtin type SwigPyPacked has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type SwigPyObject has no __module__ attribute:DeprecationWarning",
    "ignore:builtin type swigvarlink has no __module__ attribute:DeprecationWarning",
    "ignore:The 'warn' method is deprecated, use 'warning' instead:DeprecationWarning:acl_anthology.text.texmath",
    "ignore:Data directory contains a different schema.rnc as this library; you might need to update the data or the acl-anthology library.",
]
